Add in plotting code directly into analysis.
Have to do this so eventually it will have access to e.g. PCs, clusters.
Has the benefit of making me re-read and tidy analysis code.

Things to note: normalization happens first, although norm. and filtering
are independent of each other. I.e. profiles are normalized based on all
profiles, and are filtered based on all profiles. Each filtering step also
takes into account all profiles, so if CAPE removes 95% of all profiles,
and max_shear removes 75% you are not necessarily left with 25% of 5%, but
the interesection of the two sets.