Add in a new analysis step: shear_profile_analyse.
Takes on some of the work that used to be in shear_profile_plot and saves it into files.
More logical, also means that I can sign of the analysis and be sure that I'm only changing the plotting.
Have tried to reduce the dependency on pickle, shear_profile_kmeans_cluster now saves results using HDF5.
However, shear_profile_pca still needs to use it - it would be quite hard to fix this.