Use max LLS as worked out in _calc_max_low_mid_wind_diff to keep same clusters as draft2.
Do this by remapping the labels to self.remapped_labels. Doing it like this means
I can keep all other code the same, as opposed to remapping the indices, which is
a pain in the neck. There are some subleties around whether to use the map
or the argsorted map. It works at the moment but correctly reasoning about it was hard!
Basically, the map takes you from a new cluster to an old one, and vice-versa.